requirements.txt
colorama == 0.4.6
et_xmlfile == 2.0.0
numpy == 2.2.0
opencv-contrib-python == 4.10.0.84
opencv-python  == 4.10.0.84
openpyxl == 3.1.5
pip  ==  24.3.1
pycv2  == 0.0.1
pypng ==  0.20220715.0
PyQRCode == 1.2.1
pyzbar  == 0.1.9
qrcode == 8.0

